# lib/api-client-react/src/generated/api.schemas.ts/createEquipmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Print Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEquipmentRequestSpecifications } from './createEquipmentRequestSpecifications';

export interface CreateEquipmentRequest {
  groupId?: number;
  name: string;
  model?: string;
  manufacturer?: string;
  category: string;
  specifications?: CreateEquipmentRequestSpecifications;
  purchaseDate?: string;
  purchasePrice?: number;
}