# lib/api-client-react/src/generated/api.schemas.ts/createListingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Print Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateListingRequestStockType } from './createListingRequestStockType';

export interface CreateListingRequest {
  sellerId: number;
  title: string;
  description?: string | null;
  category: string;
  tags: string[];
  imageUrl?: string | null;
  basePrice: number;
  shippingCost?: number | null;
  estimatedDaysMin: number;
  estimatedDaysMax: number;
  material?: string | null;
  color?: string | null;
  productType?: string;
  equipmentUsed?: number[];
  equipmentGroups?: number[];
  isPrintOnDemand?: boolean;
  isDigitalProduct?: boolean;
  digitalFiles?: string[];
  stockType?: CreateListingRequestStockType;
}