# lib/api-client-react/src/generated/api.schemas.ts/updateEquipmentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * 3D Print Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEquipmentRequestSpecifications } from './updateEquipmentRequestSpecifications';

export interface UpdateEquipmentRequest {
  groupId?: number;
  name?: string;
  model?: string;
  manufacturer?: string;
  category?: string;
  specifications?: UpdateEquipmentRequestSpecifications;
  purchaseDate?: string;
  purchasePrice?: number;
  status?: string;
}