# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 3D Print Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings:
    get:
      operationId: listListings
      tags: [listings]
      summary: List model catalog listings
      parameters:
        - name: sellerId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: "#/components/schemas/Listing"
                  total:
                    type: integer
                required: [listings, total]
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a model listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateListingRequest"
      responses:
        "201":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
  /listings/{listingId}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing by ID
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Listing details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
    put:
      operationId: updateListing
      tags: [listings]
      summary: Update listing
      parameters:
        - name: listingId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateListingRequest"
      responses:
        "200":
          description: Updated listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Listing"
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List equipment
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment:
                    type: array
                    items:
                      $ref: "#/components/schemas/Equipment"
                required: [equipment]
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEquipmentRequest"
      responses:
        "201":
          description: Created equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{equipmentId}:
    put:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEquipmentRequest"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment deleted
components:
  schemas:
...
    Listing:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        sellerName:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        basePrice:
          type: number
        shippingCost:
          type: number
          description: Shipping charge set by seller for this listing
        estimatedDaysMin:
          type: integer
        estimatedDaysMax:
          type: integer
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        productType:
          type: string
          default: "3d_printing"
        equipmentUsed:
          type: array
          items:
            type: integer
          default: []
        equipmentGroups:
          type: array
          items:
            type: integer
          default: []
        isPrintOnDemand:
          type: boolean
          default: false
        isDigitalProduct:
          type: boolean
          default: false
        digitalFiles:
          type: array
          items:
            type: string
          default: []
        stockType:
          type: string
          enum: [inventory, print_on_demand, digital]
          default: "inventory"
        orderCount:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, sellerId, sellerName, title, category, tags, basePrice, shippingCost, estimatedDaysMin, estimatedDaysMax, orderCount, isActive, createdAt]
    CreateListingRequest:
      type: object
      properties:
        sellerId:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        basePrice:
          type: number
        shippingCost:
          type: number
          nullable: true
        estimatedDaysMin:
          type: integer
        estimatedDaysMax:
          type: integer
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        productType:
          type: string
          default: "3d_printing"
        equipmentUsed:
          type: array
          items:
            type: integer
          default: []
        equipmentGroups:
          type: array
          items:
            type: integer
          default: []
        isPrintOnDemand:
          type: boolean
          default: false
        isDigitalProduct:
          type: boolean
          default: false
        digitalFiles:
          type: array
          items:
            type: string
          default: []
        stockType:
          type: string
          enum: [inventory, print_on_demand, digital]
          default: "inventory"
      required: [sellerId, title, category, tags, basePrice, estimatedDaysMin, estimatedDaysMax]
    UpdateListingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: string
          nullable: true
        basePrice:
          type: number
        shippingCost:
          type: number
          nullable: true
        estimatedDaysMin:
          type: integer
        estimatedDaysMax:
          type: integer
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        productType:
          type: string
          default: "3d_printing"
        equipmentUsed:
          type: array
          items:
            type: integer
          default: []
        equipmentGroups:
          type: array
          items:
            type: integer
          default: []
        isPrintOnDemand:
          type: boolean
          default: false
        isDigitalProduct:
          type: boolean
          default: false
        digitalFiles:
          type: array
          items:
            type: string
          default: []
        stockType:
          type: string
          enum: [inventory, print_on_demand, digital]
          default: "inventory"
        isActive:
          type: boolean
...
    Equipment:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: integer
        groupId:
          type: integer
          nullable: true
        name:
          type: string
        model:
          type: string
          nullable: true
        manufacturer:
          type: string
          nullable: true
        category:
          type: string
        specifications:
          type: object
          nullable: true
        purchaseDate:
          type: string
          format: date-time
          nullable: true
        purchasePrice:
          type: integer
          nullable: true
        status:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sellerId, name, category, status, isActive, createdAt, updatedAt]
    CreateEquipmentRequest:
      type: object
      properties:
        groupId:
          type: integer
        name:
          type: string
        model:
          type: string
        manufacturer:
          type: string
        category:
          type: string
        specifications:
          type: object
        purchaseDate:
          type: string
          format: date-time
        purchasePrice:
          type: integer
      required: [name, category]
    UpdateEquipmentRequest:
      type: object
      properties:
        groupId:
          type: integer
        name:
          type: string
        model:
          type: string
        manufacturer:
          type: string
        category:
          type: string
        specifications:
          type: object
        purchaseDate:
          type: string
          format: date-time
        purchasePrice:
          type: integer
        status:
          type: string